debug

+++ mnist_softmax_tf2.py
@@ -58,7 +58,7 @@ start_time = timeit.default_timer()
 
 
 with sess.as_default():
-	tf.initialize_all_variables().run(feed_dict={x: batch_xs.astype('float32'), y_: batch_ys})
+	tf.initialize_all_variables().run(feed_dict={x: batch_xs.astype('float32'), y_: batch_ys.astype('float32')})
 	for j in range(50):
 		for i in range(0,50000,BATCH_SIZE):
 			# Test trained model
